# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Query Builder Service",
    "description": "Принимает декларативное JSON-описание запроса и компилирует его в Hadoop-совместимый SQL.",
    "version": "0.8.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Healthcheck",
        "description": "Проверяет доступность бекенда, фронтенда и Spark-сессии.",
        "operationId": "healthcheck_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/query/compile": {
      "post": {
        "tags": [
          "Query"
        ],
        "summary": "Эхо JSON",
        "description": "Принимает любой валидный JSON; при body={'payload': ...} используется payload (обратная совместимость).",
        "operationId": "compile_query_route_api_v1_query_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "description": "JSON object"
                  },
                  {
                    "type": "array",
                    "description": "JSON array"
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body",
                "description": "Любой JSON: object/array/string/number/bool/null. Для обратной совместимости поддерживается формат {'payload': ...}.",
                "example": {
                  "from": "table",
                  "select": [
                    "*"
                  ],
                  "orderBy": [
                    {
                      "column": "id",
                      "direction": "ASC"
                    }
                  ]
                }
              },
              "examples": {
                "object": {
                  "summary": "Объект запроса (по умолчанию)",
                  "value": {
                    "from": "table",
                    "select": [
                      "*"
                    ],
                    "orderBy": [
                      {
                        "column": "id",
                        "direction": "ASC"
                      }
                    ]
                  }
                },
                "payload": {
                  "summary": "С payload (обратная совместимость)",
                  "value": {
                    "payload": {
                      "from": "table",
                      "select": [
                        "*"
                      ],
                      "orderBy": [
                        {
                          "column": "id",
                          "direction": "ASC"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ComponentHealth": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "down"
            ],
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": "No errors"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ComponentHealth"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status"
          },
          "backend": {
            "$ref": "#/components/schemas/ComponentHealth"
          },
          "frontend": {
            "$ref": "#/components/schemas/ComponentHealth"
          },
          "spark": {
            "$ref": "#/components/schemas/ComponentHealth"
          }
        },
        "type": "object",
        "required": [
          "status",
          "backend",
          "frontend",
          "spark"
        ],
        "title": "HealthResponse"
      },
      "QueryResponse": {
        "properties": {
          "echo": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Echo",
            "description": "Эхо переданного body (или payload при формате {payload: ...})."
          },
          "sql": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql",
            "description": "Скомпилированный SQL (опционально, когда реализована компиляция)."
          },
          "columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns",
            "description": "Имена колонок результата."
          },
          "rows": {
            "anyOf": [
              {
                "items": {
                  "items": {},
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows",
            "description": "Строки результата (список списков)."
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count",
            "description": "Количество строк в ответе."
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "description": "True если результат обрезан по MAX_ROWS.",
            "default": false
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms",
            "description": "Время выполнения в мс."
          }
        },
        "type": "object",
        "title": "QueryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}